+++ fl_location/example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-// import 'dart:io';
 
 import 'package:fl_location/fl_location.dart';
 import 'package:flutter/foundation.dart';
@@ -11,109 +10,106 @@ enum ButtonState { LOADING, DONE, DISABLED }
 
 class ExampleApp extends StatefulWidget {
   @override
-  _ExampleAppState createState() => _ExampleAppState();
+  State<StatefulWidget> createState() => _ExampleAppState();
 }
 
 class _ExampleAppState extends State<ExampleApp> {
   StreamSubscription<Location>? _locationSubscription;
 
-  String _result = '';
-  ButtonState _getLocationButtonState = ButtonState.DONE;
-  ButtonState _listenLocationStreamButtonState = ButtonState.DONE;
+  final _resultText = ValueNotifier('');
+  final _getLocationButtonState = ValueNotifier(ButtonState.DONE);
+  final _subscribeLocationStreamButtonState = ValueNotifier(ButtonState.DONE);
+  final _isSubscribeLocationStream = ValueNotifier(false);
 
   Future<bool> _checkAndRequestPermission({bool? background}) async {
     if (!await FlLocation.isLocationServicesEnabled) {
-      // Location services are disabled.
+      _resultText.value = 'Location services is disabled.';
       return false;
     }
 
     var locationPermission = await FlLocation.checkLocationPermission();
     if (locationPermission == LocationPermission.deniedForever) {
-      // Cannot request runtime permission because location permission is denied forever.
+      _resultText.value = 'Location permission has been permanently denied.';
       return false;
     } else if (locationPermission == LocationPermission.denied) {
       // Ask the user for location permission.
       locationPermission = await FlLocation.requestLocationPermission();
       if (locationPermission == LocationPermission.denied ||
-          locationPermission == LocationPermission.deniedForever) return false;
+          locationPermission == LocationPermission.deniedForever) {
+        _resultText.value = 'Location permission has been denied.';
+        return false;
+      }
     }
 
-    // Location permission must always be allowed (LocationPermission.always)
+    // Location permission must always be granted (LocationPermission.always)
     // to collect location data in the background.
     if (background == true &&
-        locationPermission == LocationPermission.whileInUse) return false;
+        locationPermission == LocationPermission.whileInUse) {
+      _resultText.value =
+          'Location permission must always be granted to collect location in the background.';
+      return false;
+    }
 
-    // Location services has been enabled and permission have been granted.
     return true;
-  }
-
-  void _refreshPage() {
-    setState(() {});
   }
 
   Future<void> _getLocation() async {
     if (await _checkAndRequestPermission()) {
-      _getLocationButtonState = ButtonState.LOADING;
-      _listenLocationStreamButtonState = ButtonState.DISABLED;
-      _refreshPage();
+      _getLocationButtonState.value = ButtonState.LOADING;
+      _subscribeLocationStreamButtonState.value = ButtonState.DISABLED;
 
-      final timeLimit = const Duration(seconds: 10);
+      final Duration timeLimit = const Duration(seconds: 10);
       await FlLocation.getLocation(timeLimit: timeLimit).then((location) {
-        _result = location.toJson().toString();
-      }).onError((error, stackTrace) {
-        _result = error.toString();
+        _onLocationDetected(location);
+      }).onError((error, _) {
+        _handleError(error);
       }).whenComplete(() {
-        _getLocationButtonState = ButtonState.DONE;
-        _listenLocationStreamButtonState = ButtonState.DONE;
-        _refreshPage();
+        _getLocationButtonState.value = ButtonState.DONE;
+        _subscribeLocationStreamButtonState.value = ButtonState.DONE;
       });
     }
   }
 
-  Future<void> _listenLocationStream() async {
+  Future<void> _subscribeLocationStream() async {
     if (await _checkAndRequestPermission()) {
-      if (_locationSubscription != null) await _cancelLocationSubscription();
+      if (_locationSubscription != null) {
+        await _unsubscribeLocationStream();
+      }
 
       _locationSubscription = FlLocation.getLocationStream()
           .handleError(_handleError)
-          .listen((event) {
-        _result = event.toJson().toString();
-        _refreshPage();
-      });
+          .listen(_onLocationDetected);
 
-      _getLocationButtonState = ButtonState.DISABLED;
-      _refreshPage();
+      _getLocationButtonState.value = ButtonState.DISABLED;
+      _isSubscribeLocationStream.value = true;
     }
   }
 
-  Future<void> _cancelLocationSubscription() async {
+  Future<void> _unsubscribeLocationStream() async {
     await _locationSubscription?.cancel();
     _locationSubscription = null;
 
-    _getLocationButtonState = ButtonState.DONE;
-    _refreshPage();
+    _getLocationButtonState.value = ButtonState.DONE;
+    _isSubscribeLocationStream.value = false;
   }
 
-  void _handleError(dynamic error, StackTrace stackTrace) {
-    _result = error.toString();
-    _refreshPage();
+  void _onLocationDetected(Location location) {
+    _resultText.value = location.toJson().toString();
+  }
+
+  void _handleError(dynamic error) {
+    _resultText.value = error.toString();
   }
 
   @override
   void initState() {
     super.initState();
-    // The getLocationServicesStatusStream function is not available in Web.
+    // The getLocationServicesStatusStream function is not available on Web.
     if (!kIsWeb) {
       FlLocation.getLocationServicesStatusStream().listen((event) {
-        print('location services status: $event');
+        print('LocationServicesStatus: $event');
       });
     }
-  }
-
-  @override
-  void dispose() {
-    _locationSubscription?.cancel();
-    super.dispose();
   }
 
   @override
@@ -121,7 +117,8 @@ class _ExampleAppState extends State<ExampleApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Flutter Location'),
+          centerTitle: true,
         ),
         body: _buildContentView(),
       ),
@@ -129,36 +126,61 @@ class _ExampleAppState extends State<ExampleApp> {
   }
 
   Widget _buildContentView() {
-    final buttonList = Column(
+    final Widget buttonWidget = Column(
       children: [
         const SizedBox(height: 8.0),
-        _buildTestButton(
-          text: 'GET Location',
-          state: _getLocationButtonState,
-          onPressed: _getLocation,
+        ValueListenableBuilder(
+          valueListenable: _getLocationButtonState,
+          builder: (context, state, _) {
+            return _buildTestButton(
+              text: 'GET Location',
+              state: state,
+              onPressed: _getLocation,
+            );
+          },
         ),
         const SizedBox(height: 8.0),
-        _buildTestButton(
-          text: _locationSubscription == null
-              ? 'Listen LocationStream'
-              : 'Cancel LocationSubscription',
-          state: _listenLocationStreamButtonState,
-          onPressed: _locationSubscription == null
-              ? _listenLocationStream
-              : _cancelLocationSubscription,
+        ValueListenableBuilder(
+          valueListenable: _subscribeLocationStreamButtonState,
+          builder: (context, state, _) {
+            return ValueListenableBuilder(
+              valueListenable: _isSubscribeLocationStream,
+              builder: (context, isSubscribe, __) {
+                final String text;
+                final VoidCallback? onPressed;
+                if (isSubscribe) {
+                  text = 'Unsubscribe LocationStream';
+                  onPressed = _unsubscribeLocationStream;
+                } else {
+                  text = 'Subscribe LocationStream';
+                  onPressed = _subscribeLocationStream;
+                }
+
+                return _buildTestButton(
+                  text: text,
+                  state: state,
+                  onPressed: onPressed,
+                );
+              },
+            );
+          },
         ),
       ],
     );
 
-    final resultText =
-        Text(_result, style: Theme.of(context).textTheme.bodyText1);
+    final Widget resultWidget = ValueListenableBuilder(
+      valueListenable: _resultText,
+      builder: (context, resultText, _) {
+        return Text(resultText);
+      },
+    );
 
     return ListView(
       padding: const EdgeInsets.symmetric(horizontal: 10.0),
       children: [
-        buttonList,
+        buttonWidget,
         const SizedBox(height: 10.0),
-        resultText,
+        resultWidget,
       ],
     );
   }
@@ -166,16 +188,25 @@ class _ExampleAppState extends State<ExampleApp> {
   Widget _buildTestButton({
     required String text,
     required ButtonState state,
-    required VoidCallback? onPressed,
+    VoidCallback? onPressed,
   }) {
     return ElevatedButton(
-      child: state == ButtonState.LOADING
+      child: (state == ButtonState.LOADING)
           ? SizedBox.fromSize(
-              size: const Size(20.0, 20.0),
-              child: const CircularProgressIndicator(strokeWidth: 2.0),
-            )
+              size: const Size.square(20.0),
+              child: const CircularProgressIndicator(strokeWidth: 2.0))
           : Text(text),
-      onPressed: state == ButtonState.DONE ? onPressed : null,
+      onPressed: (state == ButtonState.DONE) ? onPressed : null,
     );
+  }
+
+  @override
+  void dispose() {
+    _locationSubscription?.cancel();
+    _resultText.dispose();
+    _getLocationButtonState.dispose();
+    _subscribeLocationStreamButtonState.dispose();
+    _isSubscribeLocationStream.dispose();
+    super.dispose();
   }
 }